Restore Python 2.6 compatibility in tests
contextlib.nested was removed in Python 3.0, because of problems
with exceptions raised in initialization of its individual context
managers. Its replacement, ExitStack, is not available in Python 2.

Emulate the old "nested" in terms of ExitStack if it's not available.

Partially reverts commit 973202b3b55fa9e8a089c1feb4db65ccc24fac35