Separate integrations to avoid dependency conflicts (ref. #86)
Right now, you cannot use the `argparse` integration without satisfying
the requirements for the `progressbar` integration as well (#86). To avoid
this problem in the future, I separated the integrations into a module per
integration. Now you can import only a specific integration:

    from bitmath.integrations.argparse import BitmathType

This change is non-invasive, as the old import command will still work
(and still lead to the error). Hence, no changes to tests are required.