Remove Python 2 types, 'long' and 'unicode'
Previously there was a conditional that mapped the old Python 2.x
long() and unicode() types to int() and str() types. That is no longer
required now that we are Python 3.x only.

In addition to the extra conditional, several type checks were able to
be removed.

Also this commit removes the click and progressbar integrations for
now. They will return later as examples in the
documentation/source. Full credit will be given to the original
authors.